# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:1323",
    "basePath": "/api/v1",
    "paths": {
        "/patients/{documentType}/{documentNumber}": {
            "get": {
                "description": "gts patient by document type and document number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Get patient",
                "operationId": "patientsGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document Type",
                        "name": "documentType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Doument Number",
                        "name": "documentNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/patient.GetResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            }
        },
        "/pharmacy/locals/{commune}": {
            "get": {
                "description": "obtiene locales al ingresarse una comuna",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Locals"
                ],
                "summary": "Get locals",
                "operationId": "getLocalsByCommune",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comuna",
                        "name": "commune",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local.GetLocalsByComunaResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.CustomError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "internalCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "local.GetLocalsByComunaResponse": {
            "type": "object",
            "properties": {
                "locals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Local"
                    }
                }
            }
        },
        "model.Local": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "model.Patient": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "documentNumber": {
                    "type": "string"
                },
                "documentType": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "secondLastName": {
                    "type": "string"
                }
            }
        },
        "patient.GetResponse": {
            "type": "object",
            "properties": {
                "patient": {
                    "$ref": "#/definitions/model.Patient"
                }
            }
        }
    }
}